feat: create contextualizer using npm script
